requirements.txt
Django==1.8
Pillow==2.9.0
boto==2.38.0
dj-static==0.0.6
django-bootstrap-form==3.2
django-bootstrap3==6.2.2
django-bootstrap3-datetimepicker==2.2.3
django-kronos==0.8
django-ses==0.7.0
django-storages==1.1.8
djangorestframework==3.2.3
gunicorn==19.3.0
kronos==0.2.2
psycopg2==2.6.1
six==1.9.0
wheel==0.24.0
wsgiref==0.1.2
